Create HillStationSafari
A Hill Station consists  of a museum and a park for safari riding. There are many passengers and n
single passenger cars.Passengers wander around the museum for a while and then line up at the park
gate to take a ride in a safari car. Note that each passenger is allowed for one ride only.
When a car is available it loads the one passenger it can hold and rides the park for a specific
amount of  time p.If all the n cars are out riding passengers around,then a passenger who wants a
ride waits. if a car is  ready to load  but there are no waiting passengers, then the cal waits.
After every r units of time One passenger from the museum  gets ready to take the safari car ride.

Assume that the park is open for k units of time. Write a program to ﬁnd the status of the passengers
and safari cars after k units of time