Create ATM
ATM Question:
On 8th November, 2016, PM of India announced Demonetisation Drive across the country. Use of old currency notes of Rs 500 and Rs 1000 is prohibited and new Rs 2000 and Rs 500 notes are introduced instead. Banks are instructed to fill all the ATMs with new notes. People usually get money from their nearby ATMs but due to unavailability of cash, people have to search for the ATMs having new notes. 



                                                                       

In north Delhi, there is a straight road which connects different areas to each other. People usually commute on this road from one area to another. Each area is represented as an integer axis and location of each area is identified with a single coordinate, which is of integer type. No two areas can be on the same coordinate. The distance between two areas will be the absolute value of difference of their coordinates, which is of integer type only. ATMs having cash to be installed in some area, but not necessarily in all of the areas. An area and the ATMs in it have the same location. For placing the new ATM, their locations must be chosen by the bank, so that the total sum of all distances between each area and its nearest ATM is minimum. You have to write a program which calculates the least possible sum of all distances between each area and its nearest new ATM. 
Input Format
Argument 1: It is the integer, which is the number of new ATMs having cash(T) to be installed. 
Argument 2: It contains array of A integers in increasing order. These A integers are the locations of the areas. For each location L, it holds that 1<= L <=10000.

Constraints
1<=T<=30
T<= A
1<= L <=10000

where number of areas is denoted by 'A',  ATMs having cash to be installed is denoted by 'T' and the coordinate of the area is denoted by 'L'.

Output Format
You need to return an integer, which is the sum of all distances between each area and its nearest new ATM. Return -1, in case of invalid arguments.  

Sample TestCase 1
Input
5
10
1
2
3
6
7
9
11
22
44
50
Output
9
Explanation
Number of areas is 10, number of new ATMs to be installed is 5 and locations of 10 areas is {1,2,3,6,7,9,11,22,44,50} as displayed in first image.








If we choose the coordinates 2, 7 , 11, 22 and 50, the sum of all distances between each area and its nearest ATM is 11 and If we choose the coordinates 2, 7 ,22, 44 and 50, sum of all distances between each area and its nearest ATM is 9. Hence, the least possible sum of all distances between each area and its nearest new ATM will be 9.